requirements.txt
paho-mqtt==1.6.1
opencv-contrib-python==4.5.3.56
numpy==1.22.3
pyyaml==6.0
tinydb==4.7.0
adafruit-io==2.6.0
python-dotenv==0.20.0